Separando los archivos de webpack
Este cambio hace que webpack sea un poquito más modular.
